+++ src/common/components/modals/DateTimePicker/datetime-picker-util.ts
@@ -1,0 +1,14 @@
+export const extractDate = (date: Date | null): string | undefined => {
+  if (!date) return undefined;
+  const year = date.getFullYear();
+  const month = String(date.getMonth() + 1).padStart(2, '0');
+  const day = String(date.getDate()).padStart(2, '0');
+  return `${year}.${month}.${day}`;
+};
+
+export const extractTime = (date: Date | null): string | undefined => {
+  if (!date) return undefined;
+  const hours = String(date.getHours()).padStart(2, '0');
+  const minutes = String(date.getMinutes()).padStart(2, '0');
+  return `${hours}:${minutes}`;
+};
